requirements.txt
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
filelock==3.4.1
Flask==2.0.3
idna==2.6
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mod-wsgi==4.9.4
oauthlib==3.2.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.6.0
typing_extensions==4.1.1
urllib3==1.22
Werkzeug==2.0.3
zipp==3.6.0
